resolve conflict

+++ server/src/main/java/com/server/fixed/service/FixedService.java
@@ -1,42 +1,38 @@
 package com.server.fixed.service;
 
-import com.server.exception.BusinessLogicException;
-import com.server.exception.ExceptionCode;
+import com.server.advice.BusinessLogicException;
+import com.server.advice.ExceptionCode;
 import com.server.fixed.entity.Fixed;
 import com.server.fixed.repository.FixedRepository;
 import com.server.trade.entity.Trade;
 import com.server.utils.CustomBeanUtils;
+import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.time.LocalDate;
+import java.util.List;
 import java.util.Optional;
 
 @Service
+@Transactional
+@RequiredArgsConstructor
 public class FixedService {
-    private final FixedRepository fixedRepository;
-    private final CustomBeanUtils<Trade> beanUtils;
 
-    public FixedService(FixedRepository fixedRepository, CustomBeanUtils<Trade> beanUtils) {
-        this.fixedRepository = fixedRepository;
-        this.beanUtils = beanUtils;
-    }
+    private final FixedRepository fixedRepository;
 
     public Fixed createFixed(Fixed fixed) {
-        Fixed saveFixed = fixedRepository.save(fixed);
-        return saveFixed;
+        return fixedRepository.save(fixed);
     }
 
-
-    public Fixed updateFixed(Fixed fixed) {
-        Fixed findFixed = findFixed(fixed.getFixedId());
-        setFixedInfos(findFixed, fixed);
-        return fixedRepository.save(findFixed);
-    }
-
-    private void setFixedInfos(Fixed findFixed, Fixed fixed) {
+    public Fixed updateFixed(Fixed fixed, Long memberId) {
+        Fixed findFixed = findVerifiedFixed(fixed.getFixedId());
+        if(!findFixed.getMemberId().equals(memberId)) {
+            throw new BusinessLogicException(ExceptionCode.FIXED_MEMBER_NOT_MATCH);
+        }
         Optional.ofNullable(fixed.getType())
                 .ifPresent(type -> findFixed.setType(type));
         Optional.ofNullable(fixed.getFixedName())
@@ -49,28 +45,38 @@ public class FixedService {
                 .ifPresent(date -> findFixed.setDate(date));
         Optional.ofNullable(fixed.getCategory())
                 .ifPresent(category -> findFixed.setCategory(category));
+        return fixedRepository.save(findFixed);
     }
 
     @Transactional(readOnly = true)
-    public Fixed findFixed(long fixedId) {
-        Fixed findFixed = fixedRepository.findById(fixedId).orElseThrow(() ->
-                new BusinessLogicException(ExceptionCode.FIXED_NOT_FOUND));
-        return findFixed;
+    public Fixed findFixed(Long fixedId, Long memberId) {
+        Fixed fixed = findVerifiedFixed(fixedId);
+        if (!fixed.getMemberId().equals(memberId)) {
+            throw new BusinessLogicException(ExceptionCode.FIXED_MEMBER_NOT_MATCH);
+        }
+        return fixed;
     }
 
-
-
     @Transactional(readOnly = true)
-    public Page<Fixed> findFixeds(int page, int size) {
-        return fixedRepository.findAll(PageRequest.of(page, size, Sort.by("date").descending()));
-    } //todo 정렬기준 고려하기
+    public List<Fixed> findFixeds(Long memberId, LocalDate startDate, LocalDate endDate) {
+        return fixedRepository.findByMemberIdAndDateBetween(memberId, startDate, endDate);
+    }
 
-    public void deleteFixed(long fixedId) {
+    public void deleteFixed(Long fixedId, Long memberId) {
+        Fixed findFixed = findVerifiedFixed(fixedId);
+        if (!findFixed.getMemberId().equals(memberId)) {
+            throw new BusinessLogicException(ExceptionCode.FIXED_MEMBER_NOT_MATCH);
+        }
         fixedRepository.deleteById(fixedId);
     }
 
-
-
-
+    // 해당 거래가 있는지 조회
+    private Fixed findVerifiedFixed(Long fixedId) {
+        Optional<Fixed> optionalFixed = fixedRepository.findById(fixedId);
+        if(optionalFixed.isEmpty()){
+            throw new BusinessLogicException(ExceptionCode.FIXED_NOT_FOUND);
+        }
+        return optionalFixed.get();
+    }
 
 }
